feat: 完善权限系统 - 移除硬编码角色名称，优化性能，支持多维度权限检查
- 移除所有硬编码的角色名称检查，改为通过权限描述符判断
- 优化权限检查性能，避免重复查询用户角色
- 完善资源访问权限检查，支持资源角色和空间隔离
- 修复用户创建 API 的错误处理逻辑
- 更新 K8s 部署文件配置格式
- 添加权限系统优化文档

+++ core/operation/resource_operation.go
@@ -33,7 +33,7 @@ func (r *ResourceOperation) CreateResourceAndAssociate(roleId int64, resourceId 
 	resourceRole := &model.ResourceRole{
 		ResourceID:   resourceId,
 		ResourceType: resourceType,
-		RoleID:       roleId,
+		RoleID:       uint(roleId),
 	}
 	// 使用 FirstOrCreate，如果已存在则更新，不存在则创建
 	var existingRole model.ResourceRole
